Vitória

+++ app/Http/Controllers/PublishersController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Publisher;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PublishersController extends Controller
 {
@@ -24,11 +26,13 @@ class PublishersController extends Controller
 
     public function create()
     {
+        Gate::authorize('bibliotecario', User::class);
         return view('publishers.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('bibliotecario', User::class);
        $validatedData = $request->validate([
            'name' => 'required|string|max:255',
            'address' => 'nullable|string|max:255',
@@ -42,6 +46,7 @@ class PublishersController extends Controller
 
     public function edit($id)
     {
+        Gate::authorize('bibliotecario', User::class);
           $publisher = Publisher::findOrFail($id);
           return view('publishers.edit', compact('publisher'));
 
@@ -49,6 +54,7 @@ class PublishersController extends Controller
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('bibliotecario', User::class);
        $validatedData = $request->validate([
                   'name' => 'required|string|max:255',
                   'address' => 'nullable|string|max:255',
@@ -63,6 +69,7 @@ class PublishersController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('bibliotecario', User::class);
        $publisher = Publisher::findOrFail($id);
                $publisher->delete();
